+++ api/src/app.ts
@@ -1,27 +1,32 @@
 import express, { Application } from "express";
+import cookieParser from "cookie-parser";
 import cors from "cors";
 import passport from "passport";
-import cookieParser from "cookie-parser";
+import { AddressInfo } from "net";
 
 import config from "./config/env.config";
 import routes from "./routes/routes";
 import errorHandler from "./middleware/errorHandler";
+import deserializeUser from "./middleware/deserializeUser";
 
 const port: number = config.port;
 const app: Application = express();
 
+// Parses form data
+app.use(cookieParser());
+app.use(express.urlencoded({ extended: true }));
+
 // Enables pre-flight across-the-board l
-app.use(cors());
+// TODO update origin for production in env.config
+// app.use(cors());
+// app.use(cors({ credentials: true, origin: "http://localhost:5173/" }));
+// app.use(cors({ credentials: true, origin: "http://127.0.0.1:5173" }));
+app.use(cors({ origin: true, credentials: true }));
 
 import "./config/passport";
-import { AddressInfo } from "net";
 
 // Parses incoming requests with JSON payloads
 app.use(express.json());
-
-// Parses form data
-app.use(express.urlencoded({ extended: true }));
-app.use(cookieParser());
 
 /* app.get("/favicon.ico", (req: Request, res: Response) => {
   res.sendStatus(204);
@@ -29,6 +34,7 @@ app.use(cookieParser());
 
 // Testing...
 app.use(passport.initialize());
+app.use(deserializeUser);
 
 // Application-level
 app.use("/", (req, res, next) => {
@@ -39,7 +45,7 @@ app.use("/", (req, res, next) => {
 // Routes
 routes(app);
 
-// No routes match
+// No paths match
 app.use((req, res, next) => {
   const error = {
     status: "fail",