+++ api/src/interfaces/user.ts
@@ -1,22 +1,22 @@
 export interface User {
   id: UserId;
-  username: Username;
-  email: string;
-  password: string;
-  role: Role;
-  timestamp: Date;
+  username?: Username;
+  email?: string;
+  password?: string;
+  role?: Role;
+  timestamp?: Date;
 }
 
 // TODO
 // For now, this fixes the error for passport.serializeUser; Express.User
-declare global {
+/* declare global {
   namespace Express {
     interface User {
-      id: UserId;
-      username: Username;
+      id?: UserId;
+      username?: Username;
     }
   }
-}
+} */
 
 type Role = "USER" | "AUTHOR" | "ADMIN";