+++ api/src/routes/auth.route.ts
@@ -3,15 +3,18 @@ import authController from "../controllers/auth.controller";
 import authenticateToken from "../middleware/authenticateToken";
 
 const authRoutes = () => {
-  const { authorize, login, signup } = authController;
+  const { authorize, refreshAccessToken, login, logout, signup } =
+    authController;
   const authRouter = Router();
 
   // GET requests
-  authRouter.get("/", authenticateToken, authorize);
+  authRouter.get("/", authorize);
+  authRouter.get("/refresh", refreshAccessToken);
 
   // POST requests
   // curl -w "\n" -X POST http://localhost:3001/auth
   authRouter.post("/login", login);
+  authRouter.post("/logout", logout);
   authRouter.post("/signup", signup);
 
   // curl -w "\n" -X POST http://localhost:3001/pizza/:id