+++ cms/src/config/axios.config.ts
@@ -3,4 +3,10 @@ import config from "./env.config";
 
 export default axios.create({
   baseURL: config.blogAPIBase,
+  withCredentials: true,
+});
+
+export const axiosPrivate = axios.create({
+  baseURL: config.blogAPIBase,
+  withCredentials: true,
 });