+++ cms/src/hooks/useAuth.tsx
@@ -1,16 +1,15 @@
 import React, {
   createContext,
+  Dispatch,
   useContext,
   useEffect,
   useMemo,
   useState,
 } from "react";
 
-import axios from "axios";
+// import axios from "axios";
+import axios from "../config/axios.config";
 
-import useLocalStorage from "./useLocalStorage";
-
-import config from "../config/env.config";
 // TODO
 // Need to set type for createContext, useState, and user
 // https://reacttraining.com/blog/react-context-with-typescript
@@ -21,6 +20,9 @@ interface AuthContext {
   login: Login;
   logout: Logout;
   accessToken: string | null;
+  setAccessToken: Dispatch<
+    string | null | ((prevState: string | null) => string | null)
+  >;
 }
 
 const AuthContext = createContext<AuthContext>({} as AuthContext);
@@ -28,10 +30,12 @@ const AuthContext = createContext<AuthContext>({} as AuthContext);
 const AuthProvider: React.FC<{ children: React.ReactNode }> = ({
   children,
 }) => {
-  const [accessToken, setAccessToken] = useState(null);
+  const [accessToken, setAccessToken] = useState<string | null>(null);
 
   const login: Login = async (newToken) => {
     console.log("login from AuthProvider running...");
+    console.log("newToken:", newToken);
+    setAccessToken(newToken);
     // setToken(newToken);
     // setIsAuthorized(true);
     // setIsLoggedOut(false);
@@ -39,36 +43,29 @@ const AuthProvider: React.FC<{ children: React.ReactNode }> = ({
 
   const logout: Logout = async () => {
     console.log("logout from AuthProvider running...");
+    setAccessToken(null);
     // setToken(null);
     // setIsAuthorized(false);
     // setIsLoggedOut(true);
   };
 
   useEffect(() => {
+    console.log("useEffect in useAuth running...");
+    console.log(accessToken);
+    // What if accessToken is null?
     const fetchToken = async () => {
-      await fetch(`${config.blogAPIBase}/auth`, {
-        method: "GET",
-        headers: {
-          authorization: `${accessToken}`,
-        },
-      }).then(async (res) => {
-        if (!res.ok) return setAccessToken(null);
-        const result = await res.json();
-        setAccessToken(result.data.access_token);
-      });
-
-      /* await axios({
-        method: "GET",
-        url: `${config.blogAPIBase}/auth`,
-        headers: { Authorization: accessToken },
-      }).then(async (res) => {
-        console.log(res);
-      }); */
+      await axios
+        .get("/auth", {
+          headers: {
+            ...(accessToken && {}),
+          },
+        })
+        .then(async (res) => {
+          console.log("res:", res);
+        });
     };
 
-    if (accessToken) {
-      fetchToken();
-    }
+    fetchToken();
   }, []);
 
   const providerValue = useMemo(() => {
@@ -76,6 +73,7 @@ const AuthProvider: React.FC<{ children: React.ReactNode }> = ({
       login,
       logout,
       accessToken,
+      setAccessToken,
     };
   }, [accessToken]);