+++ cms/src/hooks/useAxiosPrivate.tsx
@@ -1,0 +1,47 @@
+import { useEffect } from "react";
+
+import { axiosPrivate } from "../config/axios.config";
+import useRefreshToken from "./useRefreshToken";
+import { useAuth } from "./useAuth";
+
+const useAxiosPrivate = () => {
+  const refresh = useRefreshToken();
+  const { accessToken } = useAuth();
+
+  useEffect(() => {
+    const requestInterceptor = axiosPrivate.interceptors.request.use(
+      (config) => {
+        if (!config.headers["Authorization"]) {
+          config.headers["Authorization"] = `Bearer ${accessToken}`;
+        }
+
+        return config;
+      },
+      (err) => Promise.reject(err)
+    );
+
+    const responseInterceptor = axiosPrivate.interceptors.response.use(
+      (response) => response,
+      async (err) => {
+        const prevRequest = err?.config;
+
+        if (err?.response?.status === 403 && !prevRequest.sent) {
+          prevRequest.sent = true;
+          const newAccessToken = await refresh();
+          prevRequest.headers["Authorization"] = `Bearer ${newAccessToken}`;
+          return axiosPrivate(prevRequest);
+        }
+        return Promise.reject(err);
+      }
+    );
+
+    return () => {
+      axiosPrivate.interceptors.request.eject(requestInterceptor);
+      axiosPrivate.interceptors.request.eject(responseInterceptor);
+    };
+  }, [accessToken, refresh]);
+
+  return axiosPrivate;
+};
+
+export default useAxiosPrivate;